Success-2 Time Added

+++ app/src/main/java/com/project/ann/model/AdvertisedStart.kt
@@ -1,5 +1,5 @@
 package com.project.ann.model
 
 data class AdvertisedStart(
-    val seconds: Int
+    val seconds: Long
 )
